Added IDefaultCommanderProvider and 2 implementations DefinedDefaultCommanderProvider by just providing an instance. ConfiguredDefaultCommanderProvider to get the commanderType from config. Added SamplesCommander and taken the samples from (previous) MainCommander. Cleaned MainCommander having only a (forwarder) to SamplesCommander.
