+++ Core Console/appsettings.json
@@ -1,0 +1,5 @@
+{
+  //"defaultCommander": "Core_Console.MainCommander, Core Console, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
+  "defaultCommander": "Core_Console.MainCommander, Core Console"
+  //"defaultCommander": "Core_Console.SamplesCommander, Core Console"
+}
